Get rid of swiftlint warnings and move files around
